doing the same for fish events
