--- src/main/java/at/hannibal2/skyhanni/events/fishing/FishingBobberInLiquidEvent.kt
@@ -1,6 +1,0 @@
-package at.hannibal2.skyhanni.events.fishing
-
-import at.hannibal2.skyhanni.api.event.SkyHanniEvent
-import net.minecraft.world.entity.projectile.FishingHook
-
-class FishingBobberInLiquidEvent(val bobber: FishingHook, val onWater: Boolean) : SkyHanniEvent()
