--- src/main/java/at/hannibal2/skyhanni/events/fishing/FishingCatchEvent.kt
@@ -1,5 +1,0 @@
-package at.hannibal2.skyhanni.events.fishing
-
-import at.hannibal2.skyhanni.api.event.SkyHanniEvent
-
-object FishingCatchEvent : SkyHanniEvent()
