--- src/main/java/at/hannibal2/skyhanni/events/fishing/SeaCreatureFishEvent.kt
@@ -1,9 +1,0 @@
-package at.hannibal2.skyhanni.events.fishing
-
-import at.hannibal2.skyhanni.api.event.SkyHanniEvent
-import at.hannibal2.skyhanni.features.fishing.SeaCreature
-
-class SeaCreatureFishEvent(
-    val seaCreature: SeaCreature,
-    val doubleHook: Boolean,
-) : SkyHanniEvent()
