+++ src/main/java/at/hannibal2/skyhanni/features/fishing/FishingTimer.kt
@@ -10,7 +10,6 @@ import at.hannibal2.skyhanni.events.DebugDataCollectEvent
 import at.hannibal2.skyhanni.events.GuiRenderEvent
 import at.hannibal2.skyhanni.events.MobEvent
 import at.hannibal2.skyhanni.events.SecondPassedEvent
-import at.hannibal2.skyhanni.events.fishing.SeaCreatureFishEvent
 import at.hannibal2.skyhanni.events.minecraft.KeyDownEvent
 import at.hannibal2.skyhanni.features.fishing.FishingApi.babySlugName
 import at.hannibal2.skyhanni.skyhannimodule.SkyHanniModule
